Add storage breakdown chart component
